What's happening bell: if there is a notification that # should appear next to "what's happening" in the hamburger menu as well. https://3.basecamp.com/3161453/buckets/3412210/todos/620943566
